requirements.txt
Flask==3.1.1
Flask-WTF==1.2.2
WTForms==3.2.1
opencv-contrib-python-headless
tensorflow-cpu
Keras
numpy
Pillow
imutils
gunicorn
psutil>=5.9.0
requests>=2.28.0